Refactor test case package name
