+++ samza-core/src/test/java/org/apache/samza/execution/TestProcessorGraph.java
@@ -17,14 +17,12 @@
  * under the License.
  */
 
-package org.apache.samza.processorgraph;
+package org.apache.samza.execution;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import org.apache.samza.execution.ProcessorGraph;
-import org.apache.samza.execution.ProcessorNode;
 import org.apache.samza.system.StreamSpec;
 import org.junit.Before;
 import org.junit.Test;
